requirements.txt

langchain>=0.2.11
langchain-community>=0.2.10
langchain-openai>=0.1.22
langchain-google-genai>=2.0.1
langchain-postgres>=0.0.9
psycopg2-binary>=2.9.9
pgvector>=0.3.3
pypdf>=4.2.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
